Split more mixin classes out into separate files
