+++ InvenTree/plugin/base/integration/mixins.py
@@ -5,166 +5,9 @@ import logging
 
 import requests
 
-import part.models
-import stock.models
-from InvenTree.helpers import generateTestKey
-from plugin.helpers import (MixinNotImplementedError, render_template,
-                            render_text)
+from plugin.helpers import MixinNotImplementedError
 
 logger = logging.getLogger('inventree')
-
-
-class ValidationMixin:
-    """Mixin class that allows custom validation for various parts of InvenTree
-
-    Custom generation and validation functionality can be provided for:
-
-    - Part names
-    - Part IPN (internal part number) values
-    - Part parameter values
-    - Serial numbers
-    - Batch codes
-
-    Notes:
-    - Multiple ValidationMixin plugins can be used simultaneously
-    - The stub methods provided here generally return None (null value).
-    - The "first" plugin to return a non-null value for a particular method "wins"
-    - In the case of "validation" functions, all loaded plugins are checked until an exception is thrown
-
-    Implementing plugins may override any of the following methods which are of interest.
-
-    For 'validation' methods, there are three 'acceptable' outcomes:
-    - The method determines that the value is 'invalid' and raises a django.core.exceptions.ValidationError
-    - The method passes and returns None (the code then moves on to the next plugin)
-    - The method passes and returns True (and no subsequent plugins are checked)
-
-    """
-
-    class MixinMeta:
-        """Metaclass for this mixin"""
-        MIXIN_NAME = "Validation"
-
-    def __init__(self):
-        """Register the mixin"""
-        super().__init__()
-        self.add_mixin('validation', True, __class__)
-
-    def validate_part_name(self, name: str, part: part.models.Part):
-        """Perform validation on a proposed Part name
-
-        Arguments:
-            name: The proposed part name
-            part: The part instance we are validating against
-
-        Returns:
-            None or True (refer to class docstring)
-
-        Raises:
-            ValidationError if the proposed name is objectionable
-        """
-        return None
-
-    def validate_part_ipn(self, ipn: str, part: part.models.Part):
-        """Perform validation on a proposed Part IPN (internal part number)
-
-        Arguments:
-            ipn: The proposed part IPN
-            part: The Part instance we are validating against
-
-        Returns:
-            None or True (refer to class docstring)
-
-        Raises:
-            ValidationError if the proposed IPN is objectionable
-        """
-        return None
-
-    def validate_batch_code(self, batch_code: str, item: stock.models.StockItem):
-        """Validate the supplied batch code
-
-        Arguments:
-            batch_code: The proposed batch code (string)
-            item: The StockItem instance we are validating against
-
-        Returns:
-            None or True (refer to class docstring)
-
-        Raises:
-            ValidationError if the proposed batch code is objectionable
-        """
-        return None
-
-    def generate_batch_code(self):
-        """Generate a new batch code
-
-        Returns:
-            A new batch code (string) or None
-        """
-        return None
-
-    def validate_serial_number(self, serial: str, part: part.models.Part):
-        """Validate the supplied serial number.
-
-        Arguments:
-            serial: The proposed serial number (string)
-            part: The Part instance for which this serial number is being validated
-
-        Returns:
-            None or True (refer to class docstring)
-
-        Raises:
-            ValidationError if the proposed serial is objectionable
-        """
-        return None
-
-    def convert_serial_to_int(self, serial: str):
-        """Convert a serial number (string) into an integer representation.
-
-        This integer value is used for efficient sorting based on serial numbers.
-
-        A plugin which implements this method can either return:
-
-        - An integer based on the serial string, according to some algorithm
-        - A fixed value, such that serial number sorting reverts to the string representation
-        - None (null value) to let any other plugins perform the converrsion
-
-        Note that there is no requirement for the returned integer value to be unique.
-
-        Arguments:
-            serial: Serial value (string)
-
-        Returns:
-            integer representation of the serial number, or None
-        """
-        return None
-
-    def increment_serial_number(self, serial: str):
-        """Return the next sequential serial based on the provided value.
-
-        A plugin which implements this method can either return:
-
-        - A string which represents the "next" serial number in the sequence
-        - None (null value) if the next value could not be determined
-
-        Arguments:
-            serial: Current serial value (string)
-        """
-        return None
-
-    def validate_part_parameter(self, parameter, data):
-        """Validate a parameter value.
-
-        Arguments:
-            parameter: The parameter we are validating
-            data: The proposed parameter value
-
-        Returns:
-            None or True (refer to class docstring)
-
-        Raises:
-            ValidationError if the proposed parameter value is objectionable
-        """
-        pass
 
 
 class NavigationMixin:
@@ -372,141 +215,6 @@ class APICallMixin:
         if simple_response:
             return response.json()
         return response
-
-
-class PanelMixin:
-    """Mixin which allows integration of custom 'panels' into a particular page.
-
-    The mixin provides a number of key functionalities:
-
-    - Adds an (initially hidden) panel to the page
-    - Allows rendering of custom templated content to the panel
-    - Adds a menu item to the 'navbar' on the left side of the screen
-    - Allows custom javascript to be run when the panel is initially loaded
-
-    The PanelMixin class allows multiple panels to be returned for any page,
-    and also allows the plugin to return panels for many different pages.
-
-    Any class implementing this mixin must provide the 'get_custom_panels' method,
-    which dynamically returns the custom panels for a particular page.
-
-    This method is provided with:
-
-    - view : The View object which is being rendered
-    - request : The HTTPRequest object
-
-    Note that as this is called dynamically (per request),
-    then the actual panels returned can vary depending on the particular request or page
-
-    The 'get_custom_panels' method must return a list of dict objects, each with the following keys:
-
-    - title : The title of the panel, to appear in the sidebar menu
-    - description : Extra descriptive text (optional)
-    - icon : The icon to appear in the sidebar menu
-    - content : The HTML content to appear in the panel, OR
-    - content_template : A template file which will be rendered to produce the panel content
-    - javascript : The javascript content to be rendered when the panel is loade, OR
-    - javascript_template : A template file which will be rendered to produce javascript
-
-    e.g.
-
-    {
-        'title': "Updates",
-        'description': "Latest updates for this part",
-        'javascript': 'alert("You just loaded this panel!")',
-        'content': '<b>Hello world</b>',
-    }
-    """
-
-    class MixinMeta:
-        """Meta for mixin."""
-
-        MIXIN_NAME = 'Panel'
-
-    def __init__(self):
-        """Register mixin."""
-        super().__init__()
-        self.add_mixin('panel', True, __class__)
-
-    def get_custom_panels(self, view, request):
-        """This method *must* be implemented by the plugin class."""
-        raise MixinNotImplementedError(f"{__class__} is missing the 'get_custom_panels' method")
-
-    def get_panel_context(self, view, request, context):
-        """Build the context data to be used for template rendering.
-
-        Custom class can override this to provide any custom context data.
-
-        (See the example in "custom_panel_sample.py")
-        """
-        # Provide some standard context items to the template for rendering
-        context['plugin'] = self
-        context['request'] = request
-        context['user'] = getattr(request, 'user', None)
-        context['view'] = view
-
-        try:
-            context['object'] = view.get_object()
-        except AttributeError:  # pragma: no cover
-            pass
-
-        return context
-
-    def render_panels(self, view, request, context):
-        """Get panels for a view.
-
-        Args:
-            view: Current view context
-            request: Current request for passthrough
-            context: Rendering context
-
-        Returns:
-            Array of panels
-        """
-
-        panels = []
-
-        # Construct an updated context object for template rendering
-        ctx = self.get_panel_context(view, request, context)
-
-        custom_panels = self.get_custom_panels(view, request) or []
-
-        for panel in custom_panels:
-
-            content_template = panel.get('content_template', None)
-            javascript_template = panel.get('javascript_template', None)
-
-            if content_template:
-                # Render content template to HTML
-                panel['content'] = render_template(self, content_template, ctx)
-            else:
-                # Render content string to HTML
-                panel['content'] = render_text(panel.get('content', ''), ctx)
-
-            if javascript_template:
-                # Render javascript template to HTML
-                panel['javascript'] = render_template(self, javascript_template, ctx)
-            else:
-                # Render javascript string to HTML
-                panel['javascript'] = render_text(panel.get('javascript', ''), ctx)
-
-            # Check for required keys
-            required_keys = ['title', 'content']
-
-            if any(key not in panel for key in required_keys):
-                logger.warning("Custom panel for plugin %s is missing a required parameter", __class__)
-                continue
-
-            # Add some information on this plugin
-            panel['plugin'] = self
-            panel['slug'] = self.slug
-
-            # Add a 'key' for the panel, which is mostly guaranteed to be unique
-            panel['key'] = generateTestKey(self.slug + panel.get('title', 'panel'))
-
-            panels.append(panel)
-
-        return panels
 
 
 class SettingsContentMixin:
