+++ InvenTree/plugin/mixins/__init__.py
@@ -7,12 +7,13 @@ from plugin.base.barcodes.mixins import BarcodeMixin
 from plugin.base.event.mixins import EventMixin
 from plugin.base.integration.AppMixin import AppMixin
 from plugin.base.integration.mixins import (APICallMixin, NavigationMixin,
-                                            PanelMixin, SettingsContentMixin,
-                                            ValidationMixin)
+                                            SettingsContentMixin)
+from plugin.base.integration.PanelMixin import PanelMixin
 from plugin.base.integration.ReportMixin import ReportMixin
 from plugin.base.integration.ScheduleMixin import ScheduleMixin
 from plugin.base.integration.SettingsMixin import SettingsMixin
 from plugin.base.integration.UrlsMixin import UrlsMixin
+from plugin.base.integration.ValidationMixin import ValidationMixin
 from plugin.base.label.mixins import LabelPrintingMixin
 from plugin.base.locate.mixins import LocateMixin
